scenenet_pipeline/requirements.txt
addict==2.4.0
beautifulsoup4==4.11.1
cloudpickle==2.0.0
extensions==0.4
geopandas==0.12.1
h5py==3.4.0
imageio==2.9.0
ipython==8.6.0
joblib==1.0.1
laspy==2.0.3
lyft_dataset_sdk==0.0.8
matplotlib==3.4.3
mo==0.3.0
nbformat==5.1.3
numpy==1.19.5
open3d==0.10.0.0
opencv_python==4.6.0.66
openvino==2022.2.0
pandas==1.1.5
Pillow==9.3.0
pyntcloud==0.1.6
pyquaternion==0.9.9
pytest==7.2.0
PyYAML==6.0
scikit_learn==1.1.3
scipy==1.6.1
seaborn==0.11.2
setuptools==45.2.0
sympy==1.10.1
sympytorch==0.1.1
tensorflow==2.5.3
torchmetrics==0.9.0
torchvision==0.11.1+cu102
torchviz==0.0.2
tqdm==4.62.3
waymo_open_dataset==1.0.1
webcolors==1.12
yapf==0.32.0
